tests: Move TestLogBackend to backends/log.rs
