+++ tests/framework/src/backends.rs
@@ -1,5 +1,7 @@
 mod audio;
+mod log;
 mod navigator;
 
 pub use audio::TestAudioBackend;
+pub use log::TestLogBackend;
 pub use navigator::TestNavigatorBackend;
